drm_hwcomposer: fix sign-compare building error in uevent listener
d26619b5 ("drm_hwcomposer: CI: Upgrade clang-* to v12") declared 'ret' as ssize_t
but after commit 1e053b4e ("drm_hwcomposer: Make uevent listener standalone")
drm/UEventListener.cpp is affected by the following builing error:

external/drm_hwcomposer/drm/UEventListener.cpp:82:28:
error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'ssize_t' (aka 'int') [-Werror,-Wsign-compare]
    for (uint32_t i = 0; i < ret;) {
                         ~ ^ ~~~
1 error generated.

Fixes: 1e053b4e ("drm_hwcomposer: Make uevent listener standalone")
[RomanS: Fixed CI nitpicks]

+++ drm/UEventListener.cpp
@@ -79,7 +79,7 @@ void UEventListener::Routine() {
 
     bool drm_event = false;
     bool hotplug_event = false;
-    for (uint32_t i = 0; i < ret;) {
+    for (uint32_t i = 0; (ssize_t)i < ret;) {
       char *event = buffer + i;
       if (strcmp(event, "DEVTYPE=drm_minor") != 0)
         drm_event = true;
